separation of requests
Description: directory and category requests are split into two files
and the path in the router is changed. On the client side
and on the server side.
