+++ client/src/app/store/product-new/product-new.effects.ts
@@ -5,20 +5,20 @@ import { mergeMap, map, catchError } from "rxjs";
 import { ProductNewActions } from "./product-new.action";
 import { Actions, createEffect, ofType } from "@ngrx/effects";
 
-import { RequestCatalogService } from "src/app/shared/service/server/request-catalog.service";
+import { RequestCategoryService } from "src/app/shared/service/server/request-category.service";
 
 @Injectable()
 export class ProductNewEffects {
   constructor(
     private actions$: Actions,
-    private requestCatalog: RequestCatalogService
+    private requestCategory: RequestCategoryService
   ) {}
 
   getCharacteristics$ = createEffect(() =>
     this.actions$.pipe(
       ofType(ProductNewActions.getCharacteristics),
       mergeMap((actionData) => {
-        return this.requestCatalog
+        return this.requestCategory
           .getCharacteristics(actionData.categoryNumber)
           .pipe(
             map((data) => {
