+++ routes/category.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const controller = require("../controllers/category");
+const router = express.Router();
+
+router.get("/", controller.getCategory);
+router.post("/characteristics", controller.getCharacteristics);
+
+module.exports = router;
